Combine doc loader and text splitter files
