+++ libs/langchain-community/src/document_loaders/fs/oracle.ts
@@ -4,6 +4,7 @@ import path from "node:path";
 import oracledb from "oracledb";
 import * as htmlparser2 from "htmlparser2";
 import { BaseDocumentLoader } from "@langchain/core/document_loaders/base";
+import { TextSplitter, TextSplitterParams } from "@langchain/textsplitters";
 
 function* listDir(dir: string): Generator<string> {
   const files = fs.readdirSync(dir, { withFileTypes: true });
@@ -72,7 +73,7 @@ export class OracleDocLoader extends BaseDocumentLoader {
           },
         })
       );
-	  
+
       const rs = result.resultSet;
       let row;
       if (rs != null) {
@@ -143,7 +144,7 @@ export class OracleDocLoader extends BaseDocumentLoader {
           },
         })
       );
-	  
+
       const rs = result.resultSet;
       let row;
       if (rs != null) {
@@ -164,7 +165,7 @@ export class OracleDocLoader extends BaseDocumentLoader {
 
     return docs;
   }
-  
+
   // extract plain text and metadata from a row
   async _extract(row: any) {
     let plain_text = "";
@@ -173,7 +174,7 @@ export class OracleDocLoader extends BaseDocumentLoader {
     if (row != null) {
       [plain_text] = row;
       const [, html_metadata] = row;
-		
+
       const parser = new htmlparser2.Parser({
         onopentag(name, attributes) {
           if (name === "meta" && attributes.name !== undefined) {
@@ -189,3 +190,56 @@ export class OracleDocLoader extends BaseDocumentLoader {
   }
 }
 
+/**
+ * Split text into smaller pieces
+ * @example
+ * ```typescript
+ * const splitter = new OracleTextSplitter(conn, params);
+ * let chunks = await splitter.splitText(doc.pageContent);
+ * ```
+ */
+export class OracleTextSplitter extends TextSplitter {
+  protected conn: oracledb.Connection;
+
+  protected pref: Record<string, unknown>;
+
+  static lc_name() {
+    return "OracleTextSplitter";
+  }
+
+  constructor(
+    conn: oracledb.Connection,
+    pref: Record<string, unknown>,
+    fields?: TextSplitterParams
+  ) {
+    super(fields);
+    this.conn = conn;
+    this.pref = pref;
+  }
+
+  async splitText(text: string) {
+    const chunks: string[] = [];
+
+    const result = await this.conn.execute(
+      <string>(
+        `select t.column_value as data from dbms_vector_chain.utl_to_chunks(:content, json(:pref)) t`
+      ),
+      <oracledb.BindParameters>{
+        content: { val: text, dir: oracledb.BIND_IN, type: oracledb.CLOB },
+        pref: JSON.stringify(this.pref),
+      },
+      <oracledb.ExecuteOptions>(
+        (<unknown>{ fetchInfo: { DATA: { type: oracledb.STRING } } })
+      )
+    );
+    const rows: any = result.rows;
+    if (Symbol.iterator in Object(rows)) {
+      for (const row of rows) {
+        const [chunk_str] = row;
+        const chunk = JSON.parse(chunk_str);
+        chunks.push(chunk.chunk_data);
+      }
+    }
+    return chunks;
+  }
+}
